change debug flag specification
The existing `DEBUG` flag method was brittle. This change implements
a `--debug` option at the command line that can be used in/passed to
script functions as and when needed, and replicates the original
`DEBUG` method.

As a preference, the reporting handled by the `--debug` flag should
be taken care of by the chosen logger, rather than using this method,
and these changes should be considered an intermediate stopgap.